OpenAI and Ollama/Ollama Simple App/requirements.txt
langchain
langchain-community
langchain-text-splitters
python-dotenv
bs4
faiss-cpu
streamlit